fix(control-plane): restore postgres backend oauth build
Sync postgres oauth snapshot helpers with recent inventory changes and remove duplicate helper definitions so the postgres-backend feature compiles again.

+++ services/control-plane/src/store/postgres/impl_oauth_snapshot/vault_activation.rs
@@ -199,7 +199,7 @@ impl PostgresStore {
             true
         };
         let backoff = if allow_retry {
-            Some(vault_activation_backoff(next_failure_count as u32))
+            Some(vault_activation_backoff(next_failure_count))
         } else {
             None
         };
